requirements.txt
apispec==5.1.1
apispec-webframeworks==0.5.2
confluent-kafka==1.7.0
dacite==1.6.0
Flask==2.0.2
Flask-Cors==3.0.10
flask-swagger-ui==3.36.0
gunicorn==20.1.0
marshmallow==3.14.0
pycpfcnpj==1.5.1
pymongo==3.11.4
python-dotenv==0.19.1
python-keycloak==0.26.1
pytz==2022.1
Werkzeug==2.0.2
